Refactor #update_locale to its own controller
